+++ cookbook/agents/sqlite_storage.py
@@ -2,7 +2,7 @@
 
 from phi.agent import Agent
 from phi.tools.duckduckgo import DuckDuckGo
-from phi.storage.agent.sqllite import SqlAgentStorage
+from phi.storage.agent.sqlite import SqlAgentStorage
 
 agent = Agent(
     storage=SqlAgentStorage(table_name="agent_runs", db_file="data.db"),
